pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hiya-voice-agent"
version = "0.1.0"
description = "OpenAI-only LangChain voice agent (STT+LLM+TTS)"
authors = [{ name = "Hiya", email = "dev@example.com" }]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
  "langchain>=0.3.0",
  "langchain-openai>=0.2.0",
  "openai>=1.51.0",
  "httpx>=0.27.0",
  "beautifulsoup4>=4.12.0",
  "ddgs>=1.1.7",
  "sounddevice>=0.4.7",
  "numpy>=1.26",
  "webrtcvad>=2.0.10",
  "pydantic-settings>=2.4.0",
  "python-dotenv>=1.0.1",
  "tenacity>=8.3.0"
]

[project.urls]
homepage = "https://example.com/hiya-voice-agent"

[project.scripts]
hiya-agent = "hiya_voice_agent.cli:main"
hiya-web = "hiya_voice_agent.web:run"

[tool.setuptools.packages.find]
where = ["src"]


